Convert all code to be compatible with React 15.5
