+++ lib/index.test.js
@@ -2,9 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import TestUtils from 'react-addons-test-utils';
 import { expect } from 'chai';
-import Modal from '../../index';
+import Modal from '../index';
 
-/* eslint func-names:0 */
 describe('Modal', () => {
     let item;
     let modal;
